backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
